requirements.txt
asgiref==3.7.2
Django==5.0.3
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.0
Markdown==3.6
psycopg2-binary==2.9.9
sqlparse==0.4.4
